+++ src/components/ChainIconGroup.scss
@@ -46,6 +46,11 @@
   @include chain-icon
 }
 
+.sol.chain.icon {
+  background-image: url('data:image/svg+xml;utf8,<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg"><rect width="24" height="24" rx="12" fill="black"/><path d="M20.0041 16.1099L17.334 18.9306C17.2759 18.9919 17.2057 19.0407 17.1276 19.0741C17.0496 19.1075 16.9654 19.1247 16.8803 19.1246H4.22258C4.16218 19.1246 4.1031 19.1073 4.0526 19.0746C4.00209 19.042 3.96236 18.9956 3.93829 18.941C3.91421 18.8865 3.90685 18.8262 3.91709 18.7675C3.92733 18.7089 3.95474 18.6545 3.99594 18.611L6.66809 15.7903C6.72599 15.7292 6.79599 15.6805 6.87377 15.6471C6.95157 15.6137 7.03549 15.5964 7.12035 15.5963H19.7774C19.8378 15.5963 19.8969 15.6137 19.9474 15.6463C19.9979 15.6789 20.0376 15.7253 20.0618 15.7799C20.0858 15.8345 20.0932 15.8948 20.0829 15.9534C20.0727 16.012 20.0453 16.0664 20.0041 16.1099ZM17.334 10.4299C17.2759 10.3686 17.2057 10.3198 17.1276 10.2864C17.0496 10.253 16.9654 10.2358 16.8803 10.2359H4.22258C4.16218 10.2359 4.1031 10.2532 4.0526 10.2859C4.00209 10.3185 3.96236 10.3649 3.93829 10.4195C3.91421 10.474 3.90685 10.5343 3.91709 10.593C3.92733 10.6516 3.95474 10.706 3.99594 10.7495L6.66809 13.5702C6.72599 13.6313 6.79599 13.68 6.87377 13.7134C6.95157 13.7468 7.03549 13.7641 7.12035 13.7642H19.7774C19.8378 13.7642 19.8969 13.7468 19.9474 13.7142C19.9979 13.6816 20.0376 13.6352 20.0618 13.5806C20.0858 13.526 20.0932 13.4657 20.0829 13.4071C20.0727 13.3485 20.0453 13.2941 20.0041 13.2506L17.334 10.4299ZM4.22258 8.40381H16.8803C16.9654 8.40384 17.0496 8.38665 17.1276 8.35326C17.2057 8.31987 17.2759 8.27102 17.334 8.20974L20.0041 5.38908C20.0453 5.34558 20.0727 5.29118 20.0829 5.23255C20.0932 5.17392 20.0858 5.11362 20.0618 5.05906C20.0376 5.0045 19.9979 4.95806 19.9474 4.92543C19.8969 4.89281 19.8378 4.87543 19.7774 4.87543H7.12035C7.03549 4.87557 6.95157 4.89287 6.87377 4.92625C6.79599 4.95963 6.72599 5.00838 6.66809 5.06949L3.99663 7.89016C3.95547 7.9336 3.92807 7.98796 3.91781 8.04651C3.90754 8.10508 3.91485 8.16533 3.93884 8.21986C3.96283 8.2744 4.00245 8.32084 4.05285 8.35352C4.10326 8.38619 4.16224 8.40366 4.22258 8.40381Z" fill="url(%23paint0_linear_4399_445908)"/><defs><linearGradient id="paint0_linear_4399_445908" x1="5.27794" y1="19.4642" x2="18.3276" y2="4.52948" gradientUnits="userSpaceOnUse"><stop offset="0.08" stop-color="%239945FF"/><stop offset="0.3" stop-color="%238752F3"/><stop offset="0.5" stop-color="%235497D5"/><stop offset="0.6" stop-color="%2343B4CA"/><stop offset="0.72" stop-color="%2328E0B9"/><stop offset="0.97" stop-color="%2319FB9B"/></linearGradient></defs></svg>');
+  @include chain-icon
+}
+
 .random.chain.icon {
   background-color: colors.$gray-100;
   height: 24px;
