+++ src/scripts/hack-database/index.js
@@ -5,4 +5,4 @@ import '../global/cookie'
 import '../global/video'
 import '../global/index'
 // Global End
-import './table'
+import './table/index'
