+++ src/scripts/hack-database/table/togglers.ts
@@ -1,0 +1,54 @@
+const setupTogglers = (): void => {
+  const togglers = document.querySelectorAll('td[data-index]')
+  // Add event listeners to fire confetti when a button is clicked.
+  togglers.forEach((toggler) => {
+    toggler.addEventListener('click', () => {
+      toggler.classList.toggle('inverted')
+      toggleDescription((toggler as HTMLElement).dataset.index ?? '')
+    }, { passive: true })
+  })
+}
+
+const toggleDescription = (index: string): void => {
+  const descriptionRow = document.querySelector(`tr[data-index="${index}"]`)
+  const togglerRow = document.querySelector(`td[data-index="${index}"]`)
+
+  if (!(descriptionRow?.classList.contains('hidden') ?? false)) {
+    descriptionRow?.classList.add('hidden')
+    return
+  }
+
+  const descriptionRows = document.querySelectorAll('tr[data-index]')
+
+  descriptionRows.forEach((row) => {
+    row.classList.add('hidden')
+  })
+
+  const togglers = document.querySelectorAll('td[data-index]')
+
+  togglers.forEach((row) => {
+    row.classList.remove('inverted')
+  })
+
+  togglerRow?.classList.add('inverted')
+  descriptionRow?.classList.remove('hidden')
+
+  const descriptionWrapper = descriptionRow?.querySelector(
+    '.description.wrapper'
+  )
+
+  const descriptionContent = descriptionWrapper?.querySelector('.description.content');
+
+  (descriptionWrapper as HTMLElement).style.height = `${getTextNodeHeight(descriptionContent?.firstChild as Node)}px`
+}
+
+const getTextNodeHeight = (textNode: Node): number => {
+  const range = document.createRange()
+  range.selectNodeContents(textNode)
+  const rect = range.getBoundingClientRect()
+  return rect.bottom - rect.top
+}
+
+export {
+  setupTogglers
+}
