+++ src/views/hack-database/HackTable.astro
@@ -66,7 +66,7 @@ const sortableColumns = Object.keys(sortKeys);
             </span>
           </p>
         </div>
-        <div>
+        <div class="script only">
           <InputWithIcon
             className="hack search input"
             iconProps={{
