+++ util/debounce.ts
@@ -1,11 +1,11 @@
-const debounce = (func: Function, timeout?: number): Function => {
+const debounce = (func: Function, timeout: number = 300): Function => {
   let timer: NodeJS.Timeout | undefined
   return (...args: any[]) => {
     const next = (): any => func(...args)
     if (timer != null) {
       clearTimeout(timer)
     }
-    timer = setTimeout(next, (timeout ?? 0) > 0 ? timeout : 300)
+    timer = setTimeout(next, timeout)
   }
 }
